--- a/pilot/pkg/xds/bench_test.go
+++ b/pilot/pkg/xds/bench_test.go
@@ -47,8 +47,10 @@ import (
 	"istio.io/istio/pkg/config/host"
 	"istio.io/istio/pkg/config/mesh"
 	"istio.io/istio/pkg/config/schema/gvk"
+	"istio.io/istio/pkg/config/schema/kind"
 	"istio.io/istio/pkg/env"
 	istiolog "istio.io/istio/pkg/log"
+	"istio.io/istio/pkg/slices"
 	"istio.io/istio/pkg/spiffe"
 	"istio.io/istio/pkg/test"
 	"istio.io/istio/pkg/test/util/yml"
@@ -75,6 +77,8 @@ type ConfigInput struct {
 	ProxyType model.NodeType
 	// If true, all objects will be created as Kubernetes objects. Some controllers only support this mode
 	KubernetesClient bool
+	// Optionally, override the push request. Else, a full push is done
+	PushRequest *model.PushRequest
 }
 
 var testCases = []ConfigInput{
@@ -250,6 +254,43 @@ func BenchmarkSecretGeneration(b *testing.B) {
 	runBenchmark(b, v3.SecretType, secretCases)
 }
 
+var wdsCases = []ConfigInput{
+	{
+		Name:             "serviceentry-workloadentry",
+		Services:         100,
+		Instances:        1000,
+		KubernetesClient: true,
+	},
+}
+
+func BenchmarkAddressFullGeneration(b *testing.B) {
+	runBenchmark(b, v3.AddressType, wdsCases)
+}
+
+func TestAddressFullGeneration(t *testing.T) {
+	testBenchmark(t, v3.AddressType, wdsCases)
+}
+
+var wdsIncrementalCases = func() []ConfigInput {
+	cases := slices.Clone(wdsCases)
+	// Request a single resource
+	cases[0].PushRequest = &model.PushRequest{
+		ConfigsUpdated: sets.New(model.ConfigKey{
+			Kind: kind.Address,
+			Name: "Kubernetes/networking.istio.io/WorkloadEntry//random-0",
+		}),
+	}
+	return cases
+}()
+
+func BenchmarkAddressIncrementalGeneration(b *testing.B) {
+	runBenchmark(b, v3.AddressType, wdsIncrementalCases)
+}
+
+func TestAddressIncrementalGeneration(t *testing.T) {
+	testBenchmark(t, v3.AddressType, wdsIncrementalCases)
+}
+
 func createGateways(n int) map[string]*meshconfig.Network {
 	out := make(map[string]*meshconfig.Network, n)
 	for i := 0; i < n; i++ {
@@ -323,16 +364,7 @@ func runBenchmark(b *testing.B, tpe string, testCases []ConfigInput) {
 			continue
 		}
 		b.Run(tt.Name, func(b *testing.B) {
-			s, proxy := setupAndInitializeTest(b, tt)
-			wr := getWatchedResources(tpe, tt, s, proxy)
-			b.ResetTimer()
-			var c model.Resources
-			for n := 0; n < b.N; n++ {
-				c, _, _ = s.Discovery.Generators[tpe].Generate(proxy, wr, &model.PushRequest{Full: true, Push: s.PushContext()})
-				if len(c) == 0 {
-					b.Fatalf("Got no %v's!", tpe)
-				}
-			}
+			c := runBenchmarkCase(b, tt, tpe, b.N, b.ResetTimer)
 			logDebug(b, c)
 		})
 	}
@@ -351,14 +383,30 @@ func testBenchmark(t *testing.T, tpe string, testCases []ConfigInput) {
 			// No need for large test here
 			tt.Services = 1
 			tt.Instances = 1
-			s, proxy := setupAndInitializeTest(t, tt)
-			wr := getWatchedResources(tpe, tt, s, proxy)
-			c, _, _ := s.Discovery.Generators[tpe].Generate(proxy, wr, &model.PushRequest{Full: true, Push: s.PushContext()})
-			if len(c) == 0 {
-				t.Fatalf("Got no %v's!", tpe)
-			}
+			_ = runBenchmarkCase(t, tt, tpe, 1, func() {})
 		})
 	}
+}
+
+func runBenchmarkCase(t testing.TB, tt ConfigInput, tpe string, n int, reset func()) model.Resources {
+	s, proxy := setupAndInitializeTest(t, tt)
+	wr := getWatchedResources(tpe, tt, s, proxy)
+	pr := &model.PushRequest{Full: true, Push: s.PushContext()}
+	if tt.PushRequest != nil {
+		// Some types get watched resources populated on the first run. So generate with a full push first
+		_, _, _ = s.Discovery.Generators[tpe].Generate(proxy, wr, pr)
+		pr = tt.PushRequest
+		pr.Push = s.PushContext()
+	}
+	var c model.Resources
+	reset()
+	for range n {
+		c, _, _ = s.Discovery.Generators[tpe].Generate(proxy, wr, pr)
+		if len(c) == 0 {
+			t.Fatalf("Got no %v's!", tpe)
+		}
+	}
+	return c
 }
 
 func getWatchedResources(tpe string, tt ConfigInput, s *xds.FakeDiscoveryServer, proxy *model.Proxy) *model.WatchedResource {
@@ -373,6 +421,8 @@ func getWatchedResources(tpe string, tt ConfigInput, s *xds.FakeDiscoveryServer,
 		l := s.ConfigGen.BuildListeners(proxy, s.PushContext())
 		routeNames := xdstest.ExtractRoutesFromListeners(l)
 		return &model.WatchedResource{ResourceNames: sets.New(routeNames...)}
+	case v3.AddressType:
+		return &model.WatchedResource{TypeUrl: v3.AddressType, ResourceNames: sets.New[string](), Wildcard: true}
 	}
 	return nil
 }