@istio-testing 2025-01-03
@howardjohn: The following tests **failed**, say `/retest` to rerun all failed tests or `/retest-required` to rerun all mandatory failed tests:

Test name | Commit | Details | Required | Rerun command
--- | --- | --- | --- | ---
gencheck_istio | 213f75ce54e58bb256da1ad0e3b2beadf01e28eb | [link](https://prow.istio.io/view/gs/istio-prow/pr-logs/pull/istio_istio/54549/gencheck_istio/1875308389903372288) | unknown | `/test gencheck`
lint_istio | 213f75ce54e58bb256da1ad0e3b2beadf01e28eb | [link](https://prow.istio.io/view/gs/istio-prow/pr-logs/pull/istio_istio/54549/lint_istio/1875308398291980288) | unknown | `/test lint`

<details>

Instructions for interacting with me using PR comments are available [here](https://git.k8s.io/community/contributors/guide/pull-requests.md).  If you have questions or suggestions related to my behavior, please file an issue against the [kubernetes-sigs/prow](https://github.com/kubernetes-sigs/prow/issues/new?title=Prow%20issue:) repository. I understand the commands that are listed [here](https://go.k8s.io/bot-commands).
</details>
<!-- test report -->